orchestrator-agent-backend/travel_planner/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-planner"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.1.0",
    "bedrock-agentcore>=1.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = []
